requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-heroku==0.3.1
django-import-export==3.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.5.0
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
validate-email==1.3
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
